chore: display impl for localization string and reorganize
