+++ src/types/id/eoid.rs
@@ -1,0 +1,5 @@
+use crate::raw::root::RED4ext as red;
+
+#[derive(Debug, Default, Clone, Copy)]
+#[repr(transparent)]
+pub struct EditorObjectId(red::EditorObjectID);
