+++ src/types/id/guid.rs
@@ -1,0 +1,5 @@
+use crate::raw::root::RED4ext as red;
+
+#[derive(Debug, Default, Copy, Clone)]
+#[repr(transparent)]
+pub struct Guid(red::CGUID);
